+++ src/main/java/com/maddyhome/idea/vim/extension/argtextobj/VimArgTextObjExtension.java
@@ -242,7 +242,7 @@ public class VimArgTextObjExtension implements VimExtension {
     }
 
     @Override
-    public void execute(@NotNull VimEditor editor, @NotNull ExecutionContext context) {
+    public void execute(@NotNull VimEditor editor, @NotNull ExecutionContext context, @NotNull OperatorArguments operatorArguments) {
 
       IjVimEditor vimEditor = (IjVimEditor) editor;
       @NotNull VimStateMachine vimStateMachine = VimStateMachine.getInstance(vimEditor);