+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/extension/ExtensionHandler.kt
@@ -9,6 +9,7 @@ package com.maddyhome.idea.vim.extension
 
 import com.maddyhome.idea.vim.api.ExecutionContext
 import com.maddyhome.idea.vim.api.VimEditor
+import com.maddyhome.idea.vim.command.OperatorArguments
 
 /**
  * An action handler registered by its extension to be executed as a Vim command.
@@ -16,6 +17,9 @@ import com.maddyhome.idea.vim.api.VimEditor
  * @author vlan
  */
 interface ExtensionHandler {
+  @Deprecated("Please use ExtensionHandler#execute(VimEditor, ExecutionContext, OperatorArguments) instead")
+  fun execute(editor: VimEditor, context: ExecutionContext) {}
+
   /**
    * Executes the action.
    *
@@ -26,7 +30,10 @@ interface ExtensionHandler {
    * for a write action, you'll have to apply it by yourself if you're modifying IntelliJ's data structures like
    * documents or virtual files.
    */
-  fun execute(editor: VimEditor, context: ExecutionContext)
+  fun execute(editor: VimEditor, context: ExecutionContext, operatorArguments: OperatorArguments) {
+    execute(editor, context)
+  }
+
   val isRepeatable: Boolean
     get() = false